Add icon as SVG for version previous to 5.24
