+++ webapp/src/constants/svgs.js
@@ -2,5 +2,6 @@
 // See License for license information.
 
 export default {
+    VIDEO_CAMERA: "<svg width='20' height='20' viewBox='0 0 912 912' fill='none' xmlns='http://www.w3.org/2000/svg' style='vertical-align: middle;'><circle cx='456' cy='456' r='456' fill='#2D8CFF'/><path d='M199.5 349.125C199.5 333.385 212.26 320.625 228 320.625H470.25C529.275 320.625 577.125 368.475 577.125 427.5V562.875C577.125 578.615 564.365 591.375 548.625 591.375H306.375C247.35 591.375 199.5 543.525 199.5 484.5V349.125Z' fill='white'/><path fill-rule='evenodd' clip-rule='evenodd' d='M676.815 583.829L618.293 538.653C614.797 535.955 612.75 531.789 612.75 527.373V413.127C612.75 408.711 614.797 404.545 618.293 401.847L676.815 356.671C681.345 352.016 687.679 349.125 694.688 349.125C708.46 349.125 719.625 360.29 719.625 374.062V566.438C719.625 580.21 708.46 591.375 694.688 591.375C687.679 591.375 681.345 588.484 676.815 583.829Z' fill='white'/></svg>",
     VIDEO_CAMERA_3: "<svg width='19px' height='10px' viewBox='0 0 19 10' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'><title>icon video small</title><desc>Created with Sketch.</desc><defs><linearGradient x1='50%' y1='0%' x2='50%' y2='100%' id='linearGradient-1'><stop stop-color='#FFFFFF' offset='0%'></stop><stop stop-color='#F4F4F4' offset='100%'></stop></linearGradient></defs><g id='v2' stroke='none' stroke-width='1' fill='inherit' fill-rule='evenodd'><g id='zoom-bot-&amp;-drop-down' transform='translate(-443.000000, -250.000000)'><g id='zoom-bot---share-meeting' transform='translate(340.000000, 98.000000)'><g id='Send-Button' transform='translate(83.000000, 136.000000)'><g id='icon-video-small' transform='translate(20.000000, 16.000000)'><path d='M1,0 L10,0 L10,0 C12.209139,-4.05812251e-16 14,1.790861 14,4 L14,9 L14,9 C14,9.55228475 13.5522847,10 13,10 L4,10 L4,10 C1.790861,10 2.705415e-16,8.209139 0,6 L0,1 L0,1 C-6.76353751e-17,0.44771525 0.44771525,1.01453063e-16 1,0 Z' id='Rectangle-4' fill='inherit'></path><path d='M15.3667766,2.86236308 L17.3667766,1.22606168 L17.3667766,1.22606168 C17.794228,0.876342075 18.4242495,0.93935609 18.7739691,1.36680742 C18.9201398,1.54546732 19,1.76919482 19,2.00003079 L19,8.99999935 L19,8.99999935 C19,9.5522841 18.5522847,9.99999935 18,9.99999935 C17.769164,9.99999935 17.5454365,9.92013917 17.3667766,9.77396845 L15.3667766,8.13766705 L15.3667766,8.13766705 C15.1346326,7.94773826 15,7.66363761 15,7.36369795 L15,3.63633218 L15,3.63633218 C15,3.33639253 15.1346326,3.05229188 15.3667766,2.86236308 Z' id='Rectangle-5' fill='inherit'></path></g></g></g></g></g></svg>",
 };
